Make retries configurable. Move Retry logic into the retrySpec class. Move Record transformation into a separate class.
